Explore a smaller create flow
